compiling and main data grid loading

+++ Habitat/HabitatBatchBuilder.cs
@@ -133,9 +133,9 @@ namespace CHaMPWorkbench.Habitat
                     Classes.DataFolders.D50Raster(m_dD50Folder, m_sD50RasterFile, aVisit.ID, out dD50Raster);
 
                     if (theModelDef.ModelType == HabitatModelDef.ModelTypes.HSI)
-                        bSimulationRecordsOK = AddHSISimulationChildRecords(ref rSimulation, aVisit, theModelDef.ID, ref bRasterInputs, ref dD50Raster);
+                        bSimulationRecordsOK = AddHSISimulationChildRecords(ref rSimulation, aVisit, (int) theModelDef.ID, ref bRasterInputs, ref dD50Raster);
                     else
-                        bSimulationRecordsOK = AddFISSimulationChildRecords(ref rSimulation, aVisit, theModelDef.ID, ref bRasterInputs);
+                        bSimulationRecordsOK = AddFISSimulationChildRecords(ref rSimulation, aVisit, (int) theModelDef.ID, ref bRasterInputs);
 
 
                     // Need to make a quick rasterman call in order to get the rastermeta
